+++ schemars/tests/schema_name.rs
@@ -49,3 +49,29 @@ fn overriden_with_rename_multiple_type_params() -> TestResult {
         "schema-name-custom",
     )
 }
+
+#[allow(dead_code)]
+#[derive(JsonSchema)]
+#[schemars(rename = "const-generics-{BAR}-")]
+struct ConstGenericStruct<const FOO: usize, const BAR: char> {
+    foo: i32,
+}
+
+#[test]
+fn overriden_with_rename_const_generics() -> TestResult {
+    test_default_generated_schema::<ConstGenericStruct<42, 'z'>>("schema-name-const-generics")
+}
+
+#[allow(dead_code)]
+#[derive(JsonSchema)]
+struct MixedGenericStruct<T, const FOO: usize, const BAR: char> {
+    generic: T,
+    foo: i32,
+}
+
+#[test]
+fn default_name_mixed_generics() -> TestResult {
+    test_default_generated_schema::<MixedGenericStruct<MyStruct<i32, (), bool, Vec<String>>, 42, 'z'>>(
+        "schema-name-mixed-generics",
+    )
+}